refactor(cip30): lift window injection behaviour out of class
In hindsight my direction to do this inside the class (and particularly
the constructor) was wrong, when we can simply use a function and keep
the concerns of modifying window state separate. I suspect we'll end
up refactoring this further once the allow list is improved, but for
now this is an isolated change that includes more testing.