build: Add flake.nix with default package
This allows to use this package in other Nix based systems.

Note that this is focused on cardano-services, and `nix run` will call
its CLI.

The package is based on https://github.com/stephank/yarn-plugin-nixify
that is included in Yarn config and automatically generates Nix
expression that can be used for packaging in Nix.

This could be improved using isolated builds:
https://github.com/stephank/yarn-plugin-nixify/blob/main/ISOLATED_BUILDS.md
but it is incompatible with `nodeLinker: node-modules` that is set during
transition to Plug-n-Play.