fix!: convert boolean args to support ENV counterparts, parse as booleans
The Commander.js implementation currently uses boolean style options, which is incompatible with
setting ENVs to true or false, as the logic considers just the presence of it as truthy. This
change makes the options expect either 'true' or 'false', using the converter function when
parsing. `--use-queue`, `--enable-metrics`, and `--parallel` now require a value of `true` to be
provided.