fix(core): handle skipped testnet eras when configured for auto-upgrade
It's possible to configure when particular eras are upgraded, without an upgrade proposal, in
testnet cardano-node configuration, including the specification of multiple eras in the same
epoch. Era summaries therefore need to be filtered to remove eras that are being skipped,
which is evidenced by a later era starting in the same slot.