test(cardano-services): refactor chain history provider to be network agnostic
This is the work in progress for making the integration tests dynamic, so that it can be run
against any network snapshot that has sufficient information to cover all the scenarios.
Since we now have support for generating a snapshot of the e2e test suite, we can ensure control
this. It makes DB queries to set the provider arguments that were previously hard-coded.
`jest-to-match-shape-of` replaces the Jest snapshots.