fix!: free CSL resources using freeable util
The memory allocated for the CSL objects is not being managed, which results in a memory leak. Each
`coreToCsl` translator function assigns each CSL object to the scope, which is now passed in as the
first param. Management of the scope is handled by the upper scope, either via the utility function
or instantiating the scope class, then disposing of it manually. Utilities that internally use CSL
objects encapsulate their own scope.