fix(wallet): exclude submitTx reqs when determining wallet sync status
If transaction is not confirmed for maxInterval, it will be re-submitted.
If it's just taknig a long time confirm, it will get ValueNotConservedError
and then just ignore the error, assuming that transaction was already
successfully submitted.

Wallet will consider itself 'not synced' if any of the tracked requests
fail and ValueNotConservedError is technically a failed request.
Then, when the transaction is confirmed, wallet will get stuck in a
'not settled' state, as it will no longer retry to re-submit, but
stil see that 'lastRequestFailed=true'.

Including submitTx in sync status doesn't have a lot of value,
so a simple solution is to simply exclude it.

(cherry picked from commit cf8f10d15f0ca405a55352328b1b2e005f3d0985)