ci: run unit tests in nightly workflow, with vertical Node.js testing
This change covers our commitment to supporting the maintained versions of Node.js, which requires
the nightly workflow to be extended to include test runs, limited to unit tests for now. Nominating
just the major number of non-LTS versions will pick up any patch releases, but it's still important
to specify our exact LTS version, as it is set with all workflows. The last two steps are only run
with the current LTS version.