build: fix dependencies as detected by eslint import/no-extraneous-dependencies
This rule, with the configuration to specify test files, ensures packages used in `src` are listed
in the package.json `dependencies`. It will fail if missing or if the dependency is included with
`devDependencies`. Development dependencies are checked against the test directories.

There's a single disable as it's complaining about transitive dependency on `@types/pg` missing
from `dependencies`. We can see if the feature added in
https://github.com/import-js/eslint-plugin-import/pull/2543 handles this scenario once released.