test(e2e): removed browser usage outside of tests
Removed unused `ownOrigin` variable.
It was making web-extension tests fail with newer chromedriver versions (107+).
The reason was the usage of `browser` like objects outside of tests.

Unable to load spec files quite likely because they rely on `browser`
object that is not fully initialised.
Helper files that use other `browser` commands have to be moved to `before` hook