amend! feat(projection): add 'insert' and 'del' props to withStakeKeys
feat(projection)!: replace register/deregister with insert/del for withStakeKeys

This projection is intended to be used by sinks that store the current
set of registered stake keys. Those sinks are only interested in
which keys they should insert and delete, not so much in which
certificates are present in the block.