refactor: exclude collaterals from HydratedTx
collaterals field in HydratedTxBody is present only if the
transaction was successful, so it is not consumed.
In case transaction has phase 2 validation failure, collaterals
are consumed and they will not be part of the
`HydratedTxBody.collaterals` field, but will be part of `inputs`.
Optimized queries by not including collaterals as they have not
been consumed and only have historical purpose.