ci: build Dockerfile in CI without cache
It's possible for changes in package dependencies to remove the need for a local `node_modules`
directory in `cardano-services`, but our existing workflows will use available cache, which
are not invalidated by this scenario. This change ensure at least one build in CI checks we can
reliably build the image. Note that omitting the `target` tells the Docker engine to build each
stage.