refactor(cardano-services): stop throwing provider unhealthy errors from HttpService init
Change from throwing an error to logging a warning if the `HttpService` provider health is
not ok on init, to ensure it always starts, facilitating improved metrics and usability.
The original intent was that we could have assurance at startup that the service is "ready",
but that's a fairly weak situation, since the provider health could immediately revert, so
the client just gets a false sense of security. Administrators will get a warning in the logs,
and clients will now be able to use the /health endpoint to make the correct decision for
their context, with a single approach to dealing with unhealthy provider exceptions.