refactor!: simplify projection Sink to be an operator
Instead of {sink,before,after}, a Sink is now just an rxjs operator.
It gives more control to the sink implementation - it can now decide
the order of granular sinks. It is needed for stakePoolMetadata,
which requires stakePool to be run before it.

Also extract the bootstrapping part from projectIntoSink, which is
now a separate export `Bootstrap.fromCardanoNode`. It simplifies
projectIntoSink function by decoupling it form ObservableCardanoNode.
This refactor wasn't strictly required right now, but projectIntoSink
was previously too large and difficult to work with.

Lastly, remove StabilityWindowBuffer.handleEvents and only keep
tip$ and tail$, leaving the mechanism of updating the buffer up to
concrete implementations. It is no longer quired since handleEvents
is no longer called by projectIntoSink.