feat(projection-typeorm): add stakePoolMetadata sink
Includes setup of 'pg-boss', which is done by adding
a new optional 'extensions' property declared when
defining a granular sink.

Patch DataSource.initialize() to set up pg-boss schema
when any of the projections requires it. We do not need
to call PgBoss.start() during projection operation, so
it's started and stopped only once during initialization.

pg-boss schema is also extended by adding a block_height
foreign key referencing 'block' table, so rolled back jobs
are automatically deleted through ON DELETE CASCADE.

Also add a new 'dependencies' property declared when
defining a granular sink, which is used for determining
the order in which to apply the granular sinks.
stakePoolMetadata depends on stakePools, because the job
contains 'poolRegistrationId' property that represents
the pool_registration.id column