feat(projection-typeorm)!: change Block primary key col to slot
Both slot and height are unique, but slot carries more information:
it can be mapped to a timestamp and to an epoch #.

Foreign keys are typically referencing primary key and it can be
useful to have slot # in different tables - it makes it possible
to query by time or epoch without joining the block table.

BlockData (stability window buffer table) is an exception,
as it can make use of block height being sequential -
to determine # of blocks to keep, as well as to assert
data integrity.