fix(tx-construction): unneeded greedy input selection
GreedyInputSelection was used every time there were more than 1 reward accounts.
This means even if use is delegating to a single pool and has a single
registered stake key.
GreedyInputSelection should be used only when:
1. Multi delegating: to distribute the funds to multiple addresses.
2. Reducing the number of pools to 1, which implies we have multiple registered stake keys.
This is needed to concentrate the funds back into a
single address and delegate its stake key.