refactor!: remove OgmiosTxSubmitProvider
OgmiosTxSubmitProvider was unreliable as it didn't
have any re-connection logic. ObservableCardanoNode
already had re-connection logic in-place.

Furthermore, this refactor reduces the scope of Ogmios
dependency, because ogmios package is no longer Provider-aware.
This is a step towards refining this package as a thin wrapper
over ogmios client libraries.

OgmiosTxSubmitProvider is replaced by a combination of:
- ObservableCardanoNode.submitTx
- a new NodeTxSubmitProvider implementation that depends
  on ObservableCardanNode and HandleProvider

Note: re-enabling rabbitMqTxSubmitProvider and TxSubmitWorker tests
is out of scope of this commit. Those tests were already disabled.
This commit modified the test files in order for tests to compile.