refactor: remove indirect KeyAgent dependency on ObservableWallet
1. KeyAgent depends on InputResolver, which in is implemented by a util
that depends on ObservableWallet. Since ObservableWallet depends on
KeyAgent, it makes it a circular dependency. It was worked around by
having LazyWalletUtil and setupWallet utils. This commit replaces the
KeyAgent->InputResolver dependency with an extra TxContext argument for
signTransaction method, as well as removes the obsolete utils.

2. Addresses are stored by both KeyAgent and PersonalWallet.
This commit consolidates address storage in PersonalWallet,
making KeyAgent stateless and establishing single source of truth.

Notes:
- remove createAsyncKeyAgent tests instead of refactoring them,
because AsyncKeyAgent is no longer required and should be removed.
- extract address discovery and storage from PersonalWallet into
a new AddressTracker component

BREAKING CHANGE: remove KeyAgent.knownAddresses
- remove AsyncKeyAgent.knownAddresses$
- remove LazyWalletUtil and setupWallet utils
- replace KeyAgent dependency on InputResolver with props passed to sign method
- re-purpose AddressManager to Bip32Account: addresses are now stored only by the wallet