test(cardano-services): do conway proposal from account 1
This commit should be reverted once https://github.com/IntersectMBO/cardano-db-sync/issues/1702 is resolved.
Otherwise, rewards cannot be calculated correctly, which causes transactions to fail
due to not withdrawing the full amount.
This commit uses the wallet at account 1 to perform conway proposals, because this wallet is not used in other tests,
so there will not be any other transaction attempting to use the faulty rewards from proposal refund.