chore(e2e): use custom CLI
This is meant as a temporary solution until the CardanoNode embedded cardano-cli
does not have [this issue](https://github.com/IntersectMBO/cardano-cli/issues/825)
anymore, so cardano-cli 9.0.0.2 or higher.
Since cardano-cli 9.0.0.1 which has the fix cannot be used because it is dynamically
linked, and there is no newer release, I am forced to embed a customly built
cardano-cli to be used for bootstrapping the local-network.
We can remove the locally built cardano-cli when we have cli 9.0.0.2, and revert this
whole commit once it is embedded in cardano-node.