fix(wallet): preserve pouchdb collection documents when bulkDocs fails
PouchDbCollectionStore.setAll used to
1. delete all documents
2. insert the new documents

This approach has a problem that when step 2. fails, all documents from
db are gone and not re-created. For some collections such as
transactions or utxo this is not a problem, because it can recover, but
for it is a big problem for WalletRepository, because there is no way
to recover the wallets if they are lost.

This fix updates setAll to perform the following steps:
1. delete *only* the documents that are intended to be deleted
2. upsert all new documents