test(e2e): repurpose blockfrost test suite for direct from client providers
- Align "Continuous Integration - E2E BF" GHA workflow to be the same as
regular e2e, but using BlockfrostAssetProvider. We will be updating it to
use more Blockfrost providers as we refactor them to be browser compatible
- Update .env.example to be able to run Blockfrost e2e tests locally
- Simplify (reduce scope of) some assertions in nft.test.ts because,
BlockfrostAssetProvider has some fixes and returns slightly different
object. For example, it correctly does not put 'version' as otherProperties
- Temporarily disable CIP25 v2 test when running with BlockfrostAssetProvider,
there is an on-going investigation by the Blockfrost team