fix: resolve invalid CBOR serialization for maps
This is a very quick fix for #1561

Because the conditions used in `getMapSize` don't match those used in
`toCbor`, the cbor can end up invalid: for example, the map size could
be written as 8, but then only 7 values are written, if totalCollateral
is 0, for example.

A more robust solution would be to extend CborWriter to maintain a stack
of buffers, and add `startDefiniteLengthMap` and `endDefiniteLengthMap`
methods. However, we're currently blocked by this, so I did the quickest
thing possible.